+++ experimental/touchflow-cpp/src/keyboard_window.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <gtk/gtk.h>
+#include "key_injector.hpp"
+
+GtkWidget* touchflow_keyboard_window_new(KeyInjector* injector);